# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/tradeInCars": {
            "get": {
                "tags": [
                    "TradeInCar"
                ],
                "summary": "Get a listing of the TradeInCars.",
                "description": "Get all TradeInCars",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Change the Default Record Count. If not found, Returns All Records in DB.",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Change the Default Offset of the Query. If not found, 0 will be used.",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "type, 10=tradeIn; 20=evaluate",
                        "required": true,
                        "type": "integer",
                        "default": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/TradeInCar"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "TradeInCar"
                ],
                "summary": "Store a newly created TradeInCar in storage",
                "description": "Store TradeInCar",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "TradeInCar that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/TradeInCar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/TradeInCar"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/tradeInCars/{id}": {
            "get": {
                "tags": [
                    "TradeInCar"
                ],
                "summary": "Display the specified TradeInCar",
                "description": "Get TradeInCar",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of TradeInCar",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/TradeInCar"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "TradeInCar": {
            "required": [
                "owner_car_id",
                "customer_car_id",
                "amount",
                "notes"
            ],
            "properties": {
                "owner_car_id": {
                    "description": "owner_car_id",
                    "type": "integer",
                    "format": "int32"
                },
                "customer_car_id": {
                    "description": "customer_car_id",
                    "type": "integer",
                    "format": "int32"
                },
                "amount": {
                    "description": "amount",
                    "type": "number",
                    "format": "float"
                },
                "type": {
                    "description": "type, 10=tradeIn; 20=evaluate",
                    "type": "number",
                    "format": "float",
                    "default": 10
                },
                "notes": {
                    "description": "notes",
                    "type": "string"
                }
            }
        },
...
    }
}